chore: move install functions to chromedriver.js
